Extract findRequireCSSFromWithStylesCallExpression to util
I expect to need this in future rules, so I am moving it to someplace it
can be shared.
