+++ lib/rules/only-spread-css.js
@@ -1,10 +1,11 @@
 'use strict';
 
+const findRequireCSSFromWithStylesCallExpression = require('../util/findRequireCSSFromWithStylesCallExpression');
+
 module.exports = {
   meta: {
     docs: {
-      description: // eslint-disable-next-line max-len
-        'Require that css() is only spread into a JSX element without a `className` or `style` prop',
+      description: 'Require that css() is only spread into a JSX element without a `className` or `style` prop',
       recommended: true,
     },
 
@@ -20,56 +21,12 @@ module.exports = {
       return !!cssFromWithStylesName;
     }
 
-    function findRequireCSSFromWithStylesCallExpression(node) {
-      if (isCSSFound()) {
-        // We've already found it, so there is no more work to do.
-        return;
-      }
-
-      if (node.callee.name !== 'require') {
-        // foo();
-        return;
-      }
-
-      if (!node.arguments.length) {
-        // require();
-        return;
-      }
-
-      if (!node.arguments[0].value) {
-        // require(foo);
-        return;
-      }
-
-      if (!node.arguments[0].value.endsWith('withStyles')) {
-        // require('foo');
-        return;
-      }
-
-      if (node.parent.type !== 'VariableDeclarator') {
-        return;
-      }
-
-      if (node.parent.id.type !== 'ObjectPattern') {
-        // const foo = require('withStyles');
-        return;
-      }
-
-      node.parent.id.properties.some((property) => {
-        if (property.key.name !== 'css') {
-          // This is not the import that we are looking for.
-          // const { withStyles } = require('withStyles');
-          return false;
-        }
-
-        cssFromWithStylesName = property.value.name;
-        return true;
-      });
-    }
-
     return {
       CallExpression(node) {
-        findRequireCSSFromWithStylesCallExpression(node);
+        const found = findRequireCSSFromWithStylesCallExpression(node);
+        if (found !== null) {
+          cssFromWithStylesName = found;
+        }
 
         // foo()
         if (!isCSSFound()) {
